Fill NA values with zero when reading for plotting
Incomplete series can occur when there is low traffic being recorded.
Ideally the NA values would always be replaced when logging the data.
However as it currently stands the API Github provides only covers the
period where data exists, so in the long term the series could be padded
based on the date being recorded and the expected period to be covered.

In the mean time, just fix the data on the way out.